.history/requirements_20251002235159.txt
asgiref==3.8.1
channels==4.1.0
attrs==25.3.0
autobahn==24.4.2
Automat==25.4.16
cffi==1.17.1
channels_redis==4.2.1
constantly==23.10.4
cryptography==44.0.3
dj-database-url==2.3.0
Django==5.1.7
django-cors-headers==4.7.0
django-environ==0.12.0
djangorestframework==3.16.0
gunicorn==23.0.0
hyperlink==21.0.0
idna==3.10
incremental==24.7.2
msgpack==1.1.0
packaging==25.0
pillow==11.2.1
psycopg2-binary==2.9.10
pyasn1==0.6.1
pyasn1_modules==0.4.2
pycparser==2.22
pyOpenSSL==25.0.0
redis==6.0.0
service-identity==24.2.0
sqlparse==0.5.3
Twisted==24.11.0
txaio==23.1.1
typing_extensions==4.13.2
tzdata==2025.1
whitenoise==6.9.0
zope.interface==7.2
